+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 ### Integration
 
- * Breaking change to API for  `integration.HammerCTLog`:
+ * Breaking change to API for `integration.HammerCTLog`:
     * Added `ctx` as first argument, and terminate loop if it becomes cancelled
+
 ## v1.1.2
 
 ### CTFE