+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 ### Misc
  * `ctclient` tool now uses Cobra for better CLI experience (#901).
+ * #800: Removed dependency on `ratelimit`
 
 ## v1.1.2
 
@@ -28,7 +29,6 @@
 
  * Trillian from v1.3.11 to v1.4.0
  * protobuf to v2
-
 
 ## v1.1.1
 [Published 2020-10-06](https://github.com/google/certificate-transparency-go/releases/tag/v1.1.1)