+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 
  * update `google.golang.org/grpc` to v1.46.0
  * `ctclient` tool now uses Cobra for better CLI experience (#901).
+ * #800: Remove dependency from `ratelimit`.
 
 ## v1.1.2
 
@@ -32,7 +33,6 @@
 
  * Trillian from v1.3.11 to v1.4.0
  * protobuf to v2
-
 
 ## v1.1.1
 [Published 2020-10-06](https://github.com/google/certificate-transparency-go/releases/tag/v1.1.1)